Move clock out of internal
